pyproject.toml
[project]
name = "DjangoProject"
version = "0.1.0"
description = ""
authors = [
    { name = "olyaef", email = "efimovskikh@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
